+++ changelog.md
@@ -44,6 +44,8 @@
 * Fixed bug where running a limited connected equipment trace with `maximum_steps=1`
   included equipment two steps away from the starting equipment if `feeder_direction` is set.
 * Each stop condition of a traversal is now checked on each step, regardless if a previous one in the internal list has returned `True`.
+* Add `normal_upstream_trace`, `current_upstream_trace`, and `phase_inferrer` to `__all__` in `zepben.evolve.services.network.tracing.tracing`.
+* Added missing `run` method for `DownstreamTree`.
 
 ### Notes
 * None.